+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Fix issue with asset path ordering. See [#96](https://github.com/dockyard/ember-appkit-rails/issues/96) for details.
 * Fix component generator to not add \_component suffix
 * Bootstrap generator remove jbuilder
+* Don't generate views, helpers, or javascript assets with rails
+  resource generator
 
 ## 0.2.0